MODKBEKBJ-374 - Make max access types value be configurable (#213)
* MODKBEKBJ-374 - Make max access types value be configurable

* added call to mod-configuration to get the value of max access types
* added tests
* added module permission on POST and PUT /access-types request

+++ src/main/java/org/folio/repository/accesstypes/AccessTypesRepository.java
@@ -32,7 +32,7 @@ public interface AccessTypesRepository {
    */
   CompletableFuture<Void> update(String id, AccessTypeCollectionItem accessType, String tenantId);
 
-  CompletableFuture<Long> count(String tenantId);
+  CompletableFuture<Integer> count(String tenantId);
 
   /**
    * Delete access type with given id and tenantId
